model-server/requirements.txt
fastapi
uvicorn
transformers
torch
indic-nlp-library
indic-transliteration